Make last test pass
Make testRegularUserCanResetNonExistingTranscription() test
pass. It likely broke because of f614d50c.

+++ app/Controller/TranscriptionsController.php
@@ -77,7 +77,9 @@ class TranscriptionsController extends AppController
 
         $saved = false;
         if ($canEdit) {
-            $this->Transcription->delete($transcr['Transcription']['id'], false);
+            if ($transcr) {
+                $this->Transcription->delete($transcr['Transcription']['id'], false);
+            }
             $saved = $this->Transcription->generateTranscription(
                 $sentence,
                 $script,
